+++ Automatic Jigsaw puzzle solving/utility_CornerIdentificaion.h
@@ -16,7 +16,7 @@ class utility_CornerIdentificaion
 public:
 	static list<Point> identifyCorners(vector<Point> contours);
 	static list<Point> identifyTrueCorners(vector<Point> corners, Point centroid, int approxWidth);
-	static float euclideanDist(Point& p, Point& q);
+	static double euclideanDist(Point& p, Point& q);
 private:
 	static list<int> findNextPotentialTrueCorners(vector<Point> corners, Point centroid, int initialPointIndex, int approxWidth);
 };
